Add everything to a "root context" object
Create a "root context" object that contains everything else in the app, including itself.
This will make it possible to rebuild functions from source without breaking references, since everything (except for local and built-in references) will be referenced via this object. This also makes it possible for the app to "know about itself" by walking the tree.